fix: correctness/robustness pass (parsing, state, shapers, dead-code removal)
A batch of verified fixes (quality-gate green; 3540 non-browser tests pass):

Parsing:
- perl_worker_pool: drain the Perl server's stderr in a dedicated thread (a full
  stderr pipe blocked Perl mid-write and stalled parses under load), and bind
  each worker generation to its own stdout queue/reader so a dead process's
  trailing output can't leak into a restarted worker's protocol reads; do the
  process swap under the lock.
- scalar/vector: stop truncating fractional stats via int() on ingest/reduce
  (float() first, int() only as the fallback) — preserves precision.

State / shapers:
- session_repository: surface portfolio-restore failures loudly (data + per-plot
  counts, exc_info) instead of half-restoring silently.
- normalize: preserve input column order after pandas 3.0 groupby().apply().
- manager-history docstrings: 20 → 10 (match the actual rolling window).

Dead code:
- remove the Phase-10 `plot_interaction_service` shim and the dead
  `web/presenters/plot` package (callers already use the canonical paths).

UI copy: outlier-remover help now describes the actual IQR method (not "Q3").

Tests updated to match (scalar/vector precision, repository, worker pool, etc.).

+++ src/parsing/gem5/types/vector.py
@@ -171,9 +171,9 @@ class Vector(StatType):
             if not values:
                 reduced[entry] = 0
             else:
-                total = 0
+                total = 0.0
                 for i in range(self._repeat):
-                    total += int(values[i])
+                    total += float(values[i])
                 reduced[entry] = total / self._repeat
 
         object.__setattr__(self, "_reduced_content", reduced)
